added method for determining how temporal bins are made
original method used trial_time_bin_size for specific trial times. This commit includes a second option to split the epoch using a fixed integer, so if you want to split an epoch in 2 halves you'd used bin_method = fixed, n_time_bins = 2. This is helpful if epochs are of varying size.